requirements.txt
tweepy==3.10.0
pyyaml==5.1.2
mergedeep==1.3.4
boto3==1.17.52
Pillow==8.2.0
